Create reminder.jsp
This user interface helps the user recall his reminder. We would ask the user the date & time at which the user has set the reminder.Once the date & time are collected, it is checked with the database “r” and reminder time & the time entered via this interface is parsed & subtracted to obtain the time left for the user to act on the work the user set the reminder for.